Splitting test code into libraries properly
