+++ test/Bin/CMakeLists.txt
@@ -1,0 +1,4 @@
+cmake_minimum_required(VERSION 3.20.4 FATAL_ERROR)
+
+add_subdirectory(GameTest)
+add_subdirectory(UnitTest)
