+++ test/CMakeLists.txt
@@ -1,4 +1,5 @@
 cmake_minimum_required(VERSION 3.20.4 FATAL_ERROR)
 
+add_subdirectory(Bin)
 add_subdirectory(Testing)
 add_subdirectory(Tests)
