+++ test/Testing/Unit/CMakeLists.txt
@@ -1,16 +1,12 @@
 cmake_minimum_required(VERSION 3.20.4 FATAL_ERROR)
 
 if(ENABLE_TESTS)
-    set(UNIT_TEST_SOURCES UnitTestMain.cpp)
+    set(UNIT_TEST_SOURCES UnitTest.cpp)
     set(UNIT_TEST_HEADERS UnitTest.h)
 
-    add_executable(World_of_Might_and_Magic_test ${UNIT_TEST_SOURCES} ${UNIT_TEST_HEADERS})
-    target_fix_libcxx_assertions(World_of_Might_and_Magic_test)
-    target_link_libraries(World_of_Might_and_Magic_test GTest::gtest)
+    add_library(unit_test ${UNIT_TEST_SOURCES} ${UNIT_TEST_HEADERS})
+    target_link_libraries(unit_test GTest::gtest)
+    target_compile_definitions(unit_test PRIVATE TEST_GROUP=None)
 
-    add_custom_target(test World_of_Might_and_Magic_test
-        DEPENDS World_of_Might_and_Magic_test
-        WORKING_DIRECTORY ${CMAKE_CURRENT_SOURCE_DIR})
-
-    target_check_style(World_of_Might_and_Magic_test)
+    target_check_style(unit_test)
 endif()
