+++ test/Testing/Unit/UnitTest.cpp
@@ -1,0 +1,3 @@
+#include "UnitTest.h"
+
+// This source is here just to make the build system happy.
